Comment softdelete

+++ reqbaz/src/main/java/de/rwth/dbis/acis/bazaar/service/dal/DALFacadeImpl.java
@@ -496,9 +496,9 @@ public class DALFacadeImpl implements DALFacade {
     }
 
     @Override
-    public PaginationResult<Comment> listCommentsByRequirementId(int requirementId, Pageable pageable) throws BazaarException {
+    public List<Comment> listCommentsByRequirementId(int requirementId) throws BazaarException {
         commentRepository = (commentRepository != null) ? commentRepository : new CommentRepositoryImpl(dslContext);
-        return commentRepository.findAllByRequirementId(requirementId, pageable);
+        return commentRepository.findAllByRequirementId(requirementId);
     }
 
     @Override
@@ -537,7 +537,13 @@ public class DALFacadeImpl implements DALFacade {
     public Comment deleteCommentById(int commentId) throws Exception {
         commentRepository = (commentRepository != null) ? commentRepository : new CommentRepositoryImpl(dslContext);
         Comment comment = commentRepository.findById(commentId);
-        commentRepository.delete(commentId);
+        if (commentRepository.hasAnswers(commentId)) {
+            comment.setDeleted(true);
+            comment.setMessage("[This message has been deleted]");
+            commentRepository.update(comment);
+        } else{
+            commentRepository.delete(commentId);
+        }
         return comment;
     }
